+++ backend/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved router files to new `api/v1/endpoints` directory
 - Created new `core` directory for fundamental application components
 - Updated llm operations to new services in `services/llm`
+- Updated vector database operations to new services in `services/vector_db`
 
 ### Improved